+++ alicloud/validators.go
@@ -501,7 +501,7 @@ func validateAllowedStringValue(ss []string) schema.SchemaValidateFunc {
 		}
 		if !existed {
 			errors = append(errors, fmt.Errorf(
-				"%q must contain a valid string value should in array %#v, got %q",
+				"%q must contain a valid string value should be in array %#v, got %q",
 				k, ss, value))
 		}
 		return
@@ -546,7 +546,7 @@ func validateAllowedIntValue(is []int) schema.SchemaValidateFunc {
 		}
 		if !existed {
 			errors = append(errors, fmt.Errorf(
-				"%q must contain a valid int value should in array %#v, got %q",
+				"%q must contain a valid int value should be in array %#v, got %q",
 				k, is, value))
 		}
 		return
@@ -899,6 +899,9 @@ func validateJsonString(v interface{}, k string) (ws []string, errors []error) {
 	if _, err := normalizeJsonString(v); err != nil {
 		errors = append(errors, fmt.Errorf("%q contains an invalid JSON: %s", k, err))
 	}
+	if strings.Contains(v.(string), " ") || strings.Contains(v.(string), "\n") {
+		errors = append(errors, fmt.Errorf("%q can not contain any space or newline character.", k))
+	}
 	return
 }
 
@@ -1164,7 +1167,7 @@ func validateInstanceSpotStrategy(v interface{}, k string) (ws []string, errors 
 func validateDBConnectionPrefix(v interface{}, k string) (ws []string, errors []error) {
 	if value := v.(string); value != "" {
 		if len(value) < 1 || len(value) > 31 {
-			errors = append(errors, fmt.Errorf("%q cannot be less than 1 and larger than 31.", k))
+			errors = append(errors, fmt.Errorf("%q cannot be less than 1 and larger than 30.", k))
 		}
 	}
 	return
@@ -1173,7 +1176,7 @@ func validateDBConnectionPrefix(v interface{}, k string) (ws []string, errors []
 func validateDBInstanceName(v interface{}, k string) (ws []string, errors []error) {
 	if value := v.(string); value != "" {
 		if len(value) < 2 || len(value) > 256 {
-			errors = append(errors, fmt.Errorf("%q cannot be less than 2 and larger than 256.", k))
+			errors = append(errors, fmt.Errorf("%q cannot be less than 1 and larger than 30.", k))
 		}
 	}
 	return